Insert closing point for polygon rings.
D3 doesn't emit a closing point for polygon rings, so this needs to be
reinserted since it is required by GeoJSON.

If there are fewer than 3 points emitted, then we reinsert the first
point more than once, which TopoJSON also does.  Alternatively, we might
consider generating a line or a point in this situation.