Added bower support, ergo compiled JavaScript files, CONTRIBUTING doc.
Because bower requires that JavaScript codes be compiled to a single
script, rather than built with a Makefile, I have included the compiled
forms of d3-geo-projection. This appears to be the way that d3 itself is
managed on GitHub and its own bower configuration.

This also required that I change the .gitignore to not ignore those
compiled files anymore. In order to note this more formally, I have
added a CONTRIBUTING file that explains that one should not edit these
files directly.

This closes #7 pending a `bower register` by the maintainer (@mbostock)
after merging.