More precise default scales.
These scales are computed to fix almost exactly within a 959×499 area, leaving a
half-pixel margin suitable for the sphere stroke. Instead of setting a default
translate to center asymmetric projections, a default center is used instead so
that the projection can be easily scaled around the center of its bounding box.

This commit also changes the Chamberlin defaults to be more generic.