Approximate 2-d Newton-Raphson method: inverts Bertin1953, Chamberlin and interruptedSinuMollweide
The general Jacobian inversion method is detailed in Ipbuker & Bildirici 2002

In this implementation, I added a dampening factor when the determinant is too small — which helps with convergence near the singularities of the projection.

Inspired by Robert B. Schmunk’s work on the Bertin1953 inverse for G.Projector, which showed it was possible.

test all inverse projections

closes https://github.com/d3/d3-geo-projection/issues/85 (at long last)